+++ components/Reminder.py
@@ -1,5 +1,8 @@
-if __name__ != "__main__":
+if __name__ == "__main__":
+    from Database import Database
+else:
     from components.Standard import Standard
+    from components.Database import Database
 
 class Reminder():
     __Delim = "nITF0CxrEbNZa1R9Iwgpgw35jzNpk7Qpg6TZCygrZZvwa3xjN6TCo9AlWRw6XlOa1oyfY0aSAeUbk18vhD91ZwaqMh9nPD523G9M"
@@ -26,23 +29,29 @@ class Reminder():
     def ReplaceReminder(self, index, reminder):
         self.__ReminderList[index] = reminder
     
-    def SaveReminders(self):
-        result = self.__Delim.join(self.__ReminderList)
-        if __name__ == "__main__":
-            file = open("reminders.txt", "w")
-        else:
-            file = open("components/reminders.txt","w")
-        file.write(result)
-        file.close()
-    
     def LoadReminders(self):
-        if __name__ == "__main__":
-            file = open("reminders.txt", "r")
-        else:
-            file = open("components/reminders.txt","r")
-        result = file.read()
-        file.close()
-        self.__ReminderList = result.split(self.__Delim)
+        db = Database()
+        db.Get()
+        reminder = db.Data["Reminders"]
+        self.__ReminderList = reminder.split("\n")
+    
+    # def SaveReminders(self):
+    #     result = self.__Delim.join(self.__ReminderList)
+    #     if __name__ == "__main__":
+    #         file = open("reminders.txt", "w")
+    #     else:
+    #         file = open("components/reminders.txt","w")
+    #     file.write(result)
+    #     file.close()
+    
+    # def LoadReminders(self):
+    #     if __name__ == "__main__":
+    #         file = open("reminders.txt", "r")
+    #     else:
+    #         file = open("components/reminders.txt","r")
+    #     result = file.read()
+    #     file.close()
+    #     self.__ReminderList = result.split(self.__Delim)
 
 
 if __name__ == "__main__":